Added loading of maps + Dev Panel
The user can now load in a map using the File->Open menu item. This
simply allows them to find an XML file to load in. As soon as they've
picked one, it will load in the graphics, the map and render the
results.

I also added a panel overlay to put some of the Debug Buttons on.

Decided to come back to Windows Forms as WPF is causing too many
problems.