Half way to layers + other
I've added the ability to select layers base don name in the bottom
right. However currently, all rendering will be black/red. This is due
to me wanting to add a "tint" to all other layers that are not selected
to make it more visible to the user.

Other change is that the minimap is rendering a complete clone of the
larger area. Eventually, I will move each control into a class,
assosciate data with each control (view width/height) which will then
allow me to control them and then render the minimap at a higher view.