Layer Switching
I've made it so the current selected layer is move visible by dimming
the other layers and applying a little bit of red. I tried doing all
sorts before such as creating VBO, VAO and EBO to draw the tiles so I
could get the texture data out to a fragment shader then apply some red.

However in the end, it was a lot more simpler than that (as I thought it
should have been).

I also added a counter to the GameData class which was in part due to
what I was doing with the vertex data for the VBO. Anyhow, I've left it
in as I'll probably dislay it somewhere because it's some nice
information for the user.