Tiles added to panel
Tiles are now added to the panels after the graphics have loaded in. It
currently loads in one spritesheet image, loops through all the tiles
and then adds tiles that match with the same ID, otherwise it skips
them.

Each tile is a picture box with a name like "tile:0" which will be used
to parse which tile the user has clicked on.

There appears to be some issues with how they are being parsed. It's
probably down to dodgy XML files.