Can now place tiles
Although they don't save yet! Currently you must select a layer and tile
before you can begin placing the tile. Which has me thinking about ways
round it as I usually click in the window when I want to move and a tile
is always selected. So will look into a toggle.

Bug:
One bug I have found is that when navingating left or up (into negative
offset) the tile seems to want to place it's self to the right/below the
tile area you click on.