Save, Load, Icons, Middle Click
Added:
Saving of XML:
1) If loaded from XML, it will save to the same location
2) If new, it will prompt for save location
3) Save As... will clear the load location causing the user to choose
location.

Load:
1) Now asks where to load from all the time. Removed default load name.

Icons:
Added a few icons above the GL control which will be used in the future.
The icons were from the visual studio icon pack which is free to use
from Microsoft.

Middle Click:
You can now middle click on the GL window which will cause the user to
pan around the map. This is similar to functionality in 3DS Max for
example.