GL Depth + Move Layers + Toggle Ghosting
Setup the GL control to use depth and made a couple of buttons to
quickly move layers up and down. Also added a toggle for the red
ghosting.

Added a couple of shortcuts like CTRL+O, CTRL+S and F5 mainly as a quick
test but will finish the rest up later.