Undo/Redo
Currently only works with placing tiles however it should be easy to
move onto adding more now. The default drop down on the toolStrip split
buttons is awful as they seem to not have any limits on height and the
scroll doesn't work.

In the end I used code from
http://stackoverflow.com/questions/8075040/visual-studio-style-undo-drop-down-button-custom-toolstripsplitbutton
as it's a nice GUI and similar to the Visual Studio design so you know
it's going to be good!